requirements.txt

torch==2.1.0
torchvision==0.16.0
timm==0.9.7
opencv-python==4.8.0.76
albumentations==1.3.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
wandb==0.15.8
Pillow==10.0.0
kaggle==1.5.13
tqdm==4.66.1
Flask==2.3.3